ml/requirements.txt
scikit-learn
pandas
joblib
snowflake-connector-python[pandas]
python-dotenv
